Copy magic to RAM before writing to flash
Some architectures (e.g., EasyDMA on nRF52 and nRF91) cannot
DMA data directly from internal flash to an external SPI flash.
An -EIO error is returned if this is attempted.

For nRF52 with QSPI, this limitation can be avoided by using
a write buffer; i.e., CONFIG_NORDIC_QSPI_NOR_STACK_WRITE_BUFFER_SIZE
will first copy the data to a buffer in RAM before DMA'ing.

However, with the nRF91 for example, QSPI is not available and the
generic NOR SPI driver does not have a write buffer available.

This change allows writing the magic to flash, regardless of whether
or not a write buffer is enabled or available.